+++ LoTAS-Fabric/src/main/java/de/pfannekuchen/lotas/dropmanipulation/drops/package-info.java
@@ -1,0 +1,14 @@
+package de.pfannekuchen.lotas.dropmanipulation.drops;
+/** Documentation for all the subclasses
+ * 
+ *  blockdrops:
+ *   Every Class has a set of Buttons which all change the amount the Code should behave whenever a Block is being broken
+ *  
+ *  entitydrops:
+ *   Every Class has a set of Buttons which all change the amount the Code should behave whenever an Entity is being slain
+ *   
+ *  rest:
+ *   All of the rest manipulation..
+ *   
+ *   Great, isn't it?
+ */
