+++ LoTAS-Fabric/src/main/java/de/pfannekuchen/lotas/gui/ConfigurationScreen.java
@@ -14,6 +14,9 @@ import net.minecraft.client.gui.components.EditBox;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.TextComponent;
 
+/**
+ * Manages Configuration of LoTAS
+ */
 public class ConfigurationScreen extends Screen {
 
 	public ConfigurationScreen() {
