+++ LoTAS-Fabric/src/main/java/de/pfannekuchen/lotas/mixin/accessors/package-info.java
@@ -1,0 +1,8 @@
+package de.pfannekuchen.lotas.mixin.accessors;
+
+/** 
+ * An accessor is a File that allows the access of private variables in the code of minecraft, without manipulating them by injecting a method in there.
+ * 
+ * All of the following Files will do that
+ * 
+ */
