+++ LoTAS-Fabric/src/main/java/de/pfannekuchen/lotas/mixin/patches/MixinEnderPearlPatch.java
@@ -8,9 +8,16 @@ import de.pfannekuchen.lotas.core.utils.ConfigUtils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.world.item.EnderpearlItem;
 
+/**
+ * This Mixin can remove the ender pearl delay by reconnecting
+ * @author Pancake
+ */
 @Mixin(EnderpearlItem.class)
 public class MixinEnderPearlPatch {
 
+	/**
+	 * Change the amount of ticks the pearl will be delayed to
+	 */
 	@ModifyArg(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/ItemCooldowns;addCooldown(Lnet/minecraft/world/item/Item;I)V"), method = "use", index = 1)
 	public int removePearlCooldown(int ticksIn) {
 		if (ConfigUtils.getBoolean("tools", "removePearlDelay")) {
