+++ LoTAS-Fabric/src/main/java/de/pfannekuchen/lotas/mixin/patches/MixinUnbreakingPatch.java
@@ -10,9 +10,16 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import de.pfannekuchen.lotas.core.utils.ConfigUtils;
 
+/**
+ * This Mixin changes the Unbreaking Enchantment
+ * @author Pancake
+ */
 @Mixin(DigDurabilityEnchantment.class)
 public class MixinUnbreakingPatch {
 
+	/**
+	 * Modify Durability Loss Randomness
+	 */
 	@Inject(at = @At("HEAD"), method = "shouldIgnoreDurabilityDrop", cancellable = true)
 	private static void damageNO(ItemStack stack, int level, Random rand, CallbackInfoReturnable<Boolean> bool) {
 		if (ConfigUtils.getBoolean("tools", "noDamageUnbreaking") && level >= 1) {
