+++ LoTAS-Fabric/src/main/java/de/pfannekuchen/lotas/mixin/render/binds/MixinTickrateChangerSubtitleOverlay.java
@@ -7,6 +7,10 @@ import org.spongepowered.asm.mixin.injection.ModifyConstant;
 import de.pfannekuchen.lotas.mods.TickrateChangerMod;
 import net.minecraft.client.gui.components.SubtitleOverlay;
 
+/**
+ * Slows down the Subtitles
+ * @author ScribbleLP
+ */
 @Mixin(SubtitleOverlay.class)
 public abstract class MixinTickrateChangerSubtitleOverlay {
